add error handling in boats collection

+++ boats/controller.js
@@ -1,28 +1,61 @@
 import { BoatModel } from "./model.js";
 import fs from "fs/promises";
 
+// --------------------------------------------------------------------GET ALL
+
 export const getAllBoats = async (req, res) => {
   try {
+    //Wait & recibe Data
     const boats = await BoatModel.find();
-    console.log("____allBoats________⛵︎", boats);
-    res.json(boats);
+    res
+      .status(200)
+      //Confirmation back & data to frontend
+      .json({
+        success: true,
+        message: "Boats successfully retrieved ✅",
+        data: boats,
+      });
   } catch (error) {
+    // Handle errors
     console.error("Error retrieving all boats -------🦑", error);
-    res.status(500).send("Internal server error");
+    res.status(500).send({
+      success: false,
+      message: "Error retrieving all boats ❌",
+      error,
+    });
   }
 };
+
+// --------------------------------------------------------------------GET ONE
 
 export const getOneBoat = async (req, res) => {
   try {
     const { id } = req.params;
+    //Wait & recibe Data
     const boat = await BoatModel.findOne({ _id: id });
-    console.log("____getOne________⛵︎", boat);
-    res.json(boat);
+    // No Response handling
+
+    if (!dbResponse) {
+      return res.status(404).json({ message: "Boat not found" });
+    }
+    //Confirmation back  & data to frontend
+    res.status(200).json({
+      success: true,
+      message: `Movie with id= ${id} sucessfully retrieved ✅`,
+      data: boat,
+    });
   } catch (error) {
-    console.error("Error retrieving one boat -------🦑", error);
-    res.status(500).send("Internal server error");
+    // Handle errors
+    console.error("Error retrieving all boats -------🦑", error);
+    res.status(500).send({
+      success: false,
+      message: "Error retrieving one boat ❌",
+      error,
+    });
   }
 };
+
+// --------------------------------------------------------------------ADD ONE
 
 export const addOneBoat = async (req, res) => {
   try {
@@ -30,39 +63,84 @@ export const addOneBoat = async (req, res) => {
     if (req.file) {
       boat.img = req.file.path;
     }
-    console.log("____addOne________⛵︎", boat);
-    await boat.save();
-    res.end();
+    // Check if at least the boat name is present
+    if (!boat.boatname) {
+      return res.status(400).json({ error: "Boat Name is necessary." });
+    }
+    // Save the new Boot in db
+    const responseSave = await boat.save();
+
+    //Confirmation back
+    if (responseSave.acknowledged) {
+      res
+        .status(201)
+        .json({ success: true, message: "Boat successfully added ✅" })
+        .end();
+    } else {
+      // Handle Error in Database
+      res
+        .status(500)
+        .json({
+          success: false,
+          message: "Insert operation not acknowledged ❌",
+          error,
+        })
+        .end();
+    }
   } catch (error) {
     console.error("Error adding one boat -------🦑", error);
-    res.status(500).send("Internal server error");
+    res
+      .status(500)
+      .json({ success: false, message: "Error adding one boat❌", error });
   }
 };
+
+// --------------------------------------------------------------------DELETE ONE
 
 export const removeOneBoat = async (req, res) => {
   try {
     const { id } = req.params;
-    //handle delete image
+    // Save Boat to remove later the img
     const boat = await BoatModel.findOne({ _id: id });
-
+    // Remove the Boot
     const deletedBot = await BoatModel.findOneAndDelete({ _id: id });
 
+    // handle if data is not deleted
+    if (!deletedBot.deletedCount) {
+      return res
+        .status(404)
+        .json({ success: false, message: "Boat not found" });
+    }
+
+    // delete image
     if (boat) {
       await fs.unlink(boat.img);
     }
 
+    //sucess true
     console.log("____deletedOne________⛵︎", deletedBot);
-    res.json(deletedBot);
+    res.status(200).json({
+      success: true,
+      message: `Movie with id= ${id} successfully deleted ✅`,
+    });
+
     // res.end();
   } catch (error) {
-    console.error("Error removing one boat -------🦑", error);
-    res.status(500).send("Internal server error");
+    res
+      .status(500)
+      .json({ success: false, message: "Error removing one boat❌", error });
   }
 };
+
+// --------------------------------------------------------------------EDIT ONE
 
 export const editOneBoat = async (req, res) => {
   try {
     const { id } = req.params;
+
+    if (!id) {
+      return res.status(400).json({ error: "BOAT ID is missing" });
+    }
 
     //save new data & add image if it is in the request
     const newBoatData = { ...req.body };
@@ -78,14 +156,7 @@ export const editOneBoat = async (req, res) => {
     const oldData = await BoatModel.findById(id);
     const oldImage = oldData.img;
 
-    console.log("oldIMG", oldImage);
-
-    //Update with id: Mongoose assumes by default that the field that
-    //acts as the unique identifier is _id in MongoDB
-    //documents. This is a convention set by Mongoose and MongoDB.
-    // also we dont need to write _id:id
-    console.log("Boat ID:----------------", id);
-    console.log("New Boat Data:---------------", newBoatData);
+    // Update Data
 
     const updateBoat = await BoatModel.findByIdAndUpdate(id, newBoatData, {
       new: true,
@@ -102,10 +173,17 @@ export const editOneBoat = async (req, res) => {
       console.log("Image don´t change");
     }
 
-    console.log("____updateBoat________⛵︎", updateBoat);
-    res.json(updateBoat);
+    //  Confirmation back
+    res.status(201).json({
+      success: true,
+      message: `Boat with id= ${id} successfully updated ✅`,
+      data: newBoatData,
+    });
   } catch (error) {
+    // Handle errors
     console.error("Error editing one boat -------🦑", error);
-    res.status(500).send("Internal server error");
+    res
+      .status(500)
+      .json({ success: false, message: "Error editing one boat ❌", error });
   }
 };
